refactor(task-session-manager): extract board-injection subsystem
Move injected-completion handling and board append into a dedicated
board-injection.ts module. Extracted functions now take an explicit
InjectionState parameter instead of closing over closure variables,
keeping all prompt injection through the existing cache-safe-injection
helpers. index.ts drops ~207 lines (1088 -> 881).
